+++ lib/src/const.dart
@@ -2,3 +2,5 @@ const bool isProduction = bool.fromEnvironment('dart.vm.product');
 const bool isDebug = !isProduction;
 
 const packageName = 'alex';
+
+const kVerbose = 'verbose';